Added short descriptions of file contents, methods and routes to the auth and admin blueprint
