+++ devices/hmi/software/app/__init__.py
@@ -46,4 +46,8 @@ def create_app(test_config=None):
     app.register_blueprint(auths.auth)
     app.add_url_rule('/', endpoint='')
     
+    from . import admins
+    app.register_blueprint(admins.admin)
+    app.add_url_rule('/', endpoint='')
+
     return app
